DUnit Test for Row Scan Partition Pruning.
